% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{Bmi}
\alias{Bmi}
\title{Bmi
Abstract class defining the Basic Model Interface (BMI) in R.
Documentation on each function can be found at \url{https://bmi.csdms.io/}.}
\description{
Bmi
Abstract class defining the Basic Model Interface (BMI) in R.
Documentation on each function can be found at \url{https://bmi.csdms.io/}.

Bmi
Abstract class defining the Basic Model Interface (BMI) in R.
Documentation on each function can be found at \url{https://bmi.csdms.io/}.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Bmi-bmi_initialize}{\code{Bmi$bmi_initialize()}}
\item \href{#method-Bmi-update}{\code{Bmi$update()}}
\item \href{#method-Bmi-update_until}{\code{Bmi$update_until()}}
\item \href{#method-Bmi-bmi_finalize}{\code{Bmi$bmi_finalize()}}
\item \href{#method-Bmi-get_component_name}{\code{Bmi$get_component_name()}}
\item \href{#method-Bmi-get_input_item_count}{\code{Bmi$get_input_item_count()}}
\item \href{#method-Bmi-get_output_item_count}{\code{Bmi$get_output_item_count()}}
\item \href{#method-Bmi-get_input_var_names}{\code{Bmi$get_input_var_names()}}
\item \href{#method-Bmi-get_output_var_names}{\code{Bmi$get_output_var_names()}}
\item \href{#method-Bmi-get_var_grid}{\code{Bmi$get_var_grid()}}
\item \href{#method-Bmi-get_var_type}{\code{Bmi$get_var_type()}}
\item \href{#method-Bmi-get_var_units}{\code{Bmi$get_var_units()}}
\item \href{#method-Bmi-get_var_itemsize}{\code{Bmi$get_var_itemsize()}}
\item \href{#method-Bmi-get_var_nbytes}{\code{Bmi$get_var_nbytes()}}
\item \href{#method-Bmi-get_var_location}{\code{Bmi$get_var_location()}}
\item \href{#method-Bmi-get_current_time}{\code{Bmi$get_current_time()}}
\item \href{#method-Bmi-get_start_time}{\code{Bmi$get_start_time()}}
\item \href{#method-Bmi-get_end_time}{\code{Bmi$get_end_time()}}
\item \href{#method-Bmi-get_time_units}{\code{Bmi$get_time_units()}}
\item \href{#method-Bmi-get_time_step}{\code{Bmi$get_time_step()}}
\item \href{#method-Bmi-get_value}{\code{Bmi$get_value()}}
\item \href{#method-Bmi-get_value_at_indices}{\code{Bmi$get_value_at_indices()}}
\item \href{#method-Bmi-get_value_ptr}{\code{Bmi$get_value_ptr()}}
\item \href{#method-Bmi-set_value}{\code{Bmi$set_value()}}
\item \href{#method-Bmi-set_value_at_indices}{\code{Bmi$set_value_at_indices()}}
\item \href{#method-Bmi-get_grid_rank}{\code{Bmi$get_grid_rank()}}
\item \href{#method-Bmi-get_grid_size}{\code{Bmi$get_grid_size()}}
\item \href{#method-Bmi-get_grid_type}{\code{Bmi$get_grid_type()}}
\item \href{#method-Bmi-get_grid_shape}{\code{Bmi$get_grid_shape()}}
\item \href{#method-Bmi-get_grid_spacing}{\code{Bmi$get_grid_spacing()}}
\item \href{#method-Bmi-get_grid_origin}{\code{Bmi$get_grid_origin()}}
\item \href{#method-Bmi-get_grid_x}{\code{Bmi$get_grid_x()}}
\item \href{#method-Bmi-get_grid_y}{\code{Bmi$get_grid_y()}}
\item \href{#method-Bmi-get_grid_z}{\code{Bmi$get_grid_z()}}
\item \href{#method-Bmi-get_grid_node_count}{\code{Bmi$get_grid_node_count()}}
\item \href{#method-Bmi-get_grid_edge_count}{\code{Bmi$get_grid_edge_count()}}
\item \href{#method-Bmi-get_grid_face_count}{\code{Bmi$get_grid_face_count()}}
\item \href{#method-Bmi-get_grid_edge_nodes}{\code{Bmi$get_grid_edge_nodes()}}
\item \href{#method-Bmi-get_grid_face_nodes}{\code{Bmi$get_grid_face_nodes()}}
\item \href{#method-Bmi-get_grid_face_edges}{\code{Bmi$get_grid_face_edges()}}
\item \href{#method-Bmi-get_grid_nodes_per_face}{\code{Bmi$get_grid_nodes_per_face()}}
\item \href{#method-Bmi-clone}{\code{Bmi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-bmi_initialize"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-bmi_initialize}{}}}
\subsection{Method \code{bmi_initialize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$bmi_initialize(config_file)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-update"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$update()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-update_until"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-update_until}{}}}
\subsection{Method \code{update_until()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$update_until(time)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-bmi_finalize"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-bmi_finalize}{}}}
\subsection{Method \code{bmi_finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$bmi_finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-get_component_name"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-get_component_name}{}}}
\subsection{Method \code{get_component_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$get_component_name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-get_input_item_count"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-get_input_item_count}{}}}
\subsection{Method \code{get_input_item_count()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$get_input_item_count()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-get_output_item_count"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-get_output_item_count}{}}}
\subsection{Method \code{get_output_item_count()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$get_output_item_count()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-get_input_var_names"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-get_input_var_names}{}}}
\subsection{Method \code{get_input_var_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$get_input_var_names()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-get_output_var_names"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-get_output_var_names}{}}}
\subsection{Method \code{get_output_var_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$get_output_var_names()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-get_var_grid"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-get_var_grid}{}}}
\subsection{Method \code{get_var_grid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$get_var_grid(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-get_var_type"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-get_var_type}{}}}
\subsection{Method \code{get_var_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$get_var_type(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-get_var_units"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-get_var_units}{}}}
\subsection{Method \code{get_var_units()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$get_var_units(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-get_var_itemsize"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-get_var_itemsize}{}}}
\subsection{Method \code{get_var_itemsize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$get_var_itemsize(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-get_var_nbytes"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-get_var_nbytes}{}}}
\subsection{Method \code{get_var_nbytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$get_var_nbytes(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-get_var_location"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-get_var_location}{}}}
\subsection{Method \code{get_var_location()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$get_var_location(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-get_current_time"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-get_current_time}{}}}
\subsection{Method \code{get_current_time()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$get_current_time()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-get_start_time"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-get_start_time}{}}}
\subsection{Method \code{get_start_time()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$get_start_time()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-get_end_time"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-get_end_time}{}}}
\subsection{Method \code{get_end_time()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$get_end_time()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-get_time_units"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-get_time_units}{}}}
\subsection{Method \code{get_time_units()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$get_time_units()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-get_time_step"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-get_time_step}{}}}
\subsection{Method \code{get_time_step()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$get_time_step()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-get_value"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-get_value}{}}}
\subsection{Method \code{get_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$get_value(name, dest)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-get_value_at_indices"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-get_value_at_indices}{}}}
\subsection{Method \code{get_value_at_indices()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$get_value_at_indices(name, inds, dest)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-get_value_ptr"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-get_value_ptr}{}}}
\subsection{Method \code{get_value_ptr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$get_value_ptr(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-set_value"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-set_value}{}}}
\subsection{Method \code{set_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$set_value(name, src)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-set_value_at_indices"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-set_value_at_indices}{}}}
\subsection{Method \code{set_value_at_indices()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$set_value_at_indices(name, inds, src)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-get_grid_rank"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-get_grid_rank}{}}}
\subsection{Method \code{get_grid_rank()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$get_grid_rank(grid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-get_grid_size"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-get_grid_size}{}}}
\subsection{Method \code{get_grid_size()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$get_grid_size(grid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-get_grid_type"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-get_grid_type}{}}}
\subsection{Method \code{get_grid_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$get_grid_type(grid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-get_grid_shape"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-get_grid_shape}{}}}
\subsection{Method \code{get_grid_shape()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$get_grid_shape(grid, shape)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-get_grid_spacing"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-get_grid_spacing}{}}}
\subsection{Method \code{get_grid_spacing()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$get_grid_spacing(grid, spacing)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-get_grid_origin"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-get_grid_origin}{}}}
\subsection{Method \code{get_grid_origin()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$get_grid_origin(grid, origin)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-get_grid_x"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-get_grid_x}{}}}
\subsection{Method \code{get_grid_x()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$get_grid_x(grid, x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-get_grid_y"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-get_grid_y}{}}}
\subsection{Method \code{get_grid_y()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$get_grid_y(grid, y)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-get_grid_z"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-get_grid_z}{}}}
\subsection{Method \code{get_grid_z()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$get_grid_z(grid, z)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-get_grid_node_count"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-get_grid_node_count}{}}}
\subsection{Method \code{get_grid_node_count()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$get_grid_node_count(grid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-get_grid_edge_count"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-get_grid_edge_count}{}}}
\subsection{Method \code{get_grid_edge_count()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$get_grid_edge_count(grid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-get_grid_face_count"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-get_grid_face_count}{}}}
\subsection{Method \code{get_grid_face_count()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$get_grid_face_count(grid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-get_grid_edge_nodes"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-get_grid_edge_nodes}{}}}
\subsection{Method \code{get_grid_edge_nodes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$get_grid_edge_nodes(grid, edge_nodes)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-get_grid_face_nodes"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-get_grid_face_nodes}{}}}
\subsection{Method \code{get_grid_face_nodes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$get_grid_face_nodes(grid, face_nodes)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-get_grid_face_edges"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-get_grid_face_edges}{}}}
\subsection{Method \code{get_grid_face_edges()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$get_grid_face_edges(grid, face_edges)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-get_grid_nodes_per_face"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-get_grid_nodes_per_face}{}}}
\subsection{Method \code{get_grid_nodes_per_face()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$get_grid_nodes_per_face(grid, nodes_per_face)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bmi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Bmi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bmi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
